Move ...azure.vmagent tests into same package as code under test.
